requirements.txt
cat > requirements.txt << 'EOF'
streamlit
python-docx
pandas
EOF
